move scraping to utilities
